fix(auth): Ensuring test user account cleanup with a Rule
